Basic Styles and File Reading
SCSS has been added to make the page a little more user-friendly. The logic.js file was added in order to operate on the front end with the file reading and displaying. For now, the content read is passed onto a run function, which will process the data. In case of errors, a showError function will receive an errorMsg string from run, which will be displayed to the user.