Styling and Basic Heap Manipulation
The interface is now better organized, and using the command 'exibe' displays the heap, the line where the display command was executed and the method that was being used at that moment. Logic.js now has the ability to interpret basic vocabulary: heap (useless for now), new (working, for the current default heap first), del (working), exibe (working), and  assignments (experimental: not yet implemented in heap structure, but tested with a separate vars object). Not all errors are handled accordingly, for now.