Major Changes
Heap manipulation is now handled differently. An array describes the available space and, according to it, the insertion function knows if it's possible to insert variables at specific locations in the heap. A separate object stores the variable names and their associated information. First and Next methods have been implemented.